[ 機能追加 ] 推奨外バージョンのPHP使用時のプラグイン無効化 #192
